+++ db/migrate/20221115101700_add_organization_limits.rb
@@ -1,0 +1,6 @@
+class AddOrganizationLimits < ActiveRecord::Migration[5.2]
+  def change
+    add_column :organizations, :server_limit, :integer
+    add_column :organizations, :server_send_limit, :integer
+  end
+end